Refactor
Getting ready for multiple render types
